pyproject.toml
[tool.poetry]
name = "zrive-ds"
version = "0.1.0"
description = "Project for the program Zrive Applied Data Science"
authors = ["Guille Barquero <gb@zriveapp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
numpy = "^1.25.2"
requests = "^2.31.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pytest = "^7.4.0"
scikit-learn = "^1.4.1.post1"
python-dotenv = "^1.1.0"
boto3 = "^1.38.8"
pyarrow = "^20.0.0"
notebook = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
